buggen fixad?

+++ BookStore/Controllers/CategoryController.cs
@@ -3,11 +3,13 @@
 using BookStore.DbAccess;
 using BookStore.Models;
 using Microsoft.AspNetCore.Mvc;
+using BookStore.Authorize;
 //using System.Web.Http;
 
 [Route("api/[controller]")]
 //[Route("api/[controller]")]
 [ApiController]
+[Authorize]
 public class CategoryController : ControllerBase
 {
     private readonly CategoryCrud _categoryCrud;
@@ -18,28 +20,30 @@ public class CategoryController : ControllerBase
 
     // GET: api/<CategoryController>
     [HttpGet]
+    [AllowAnonymous]
     public async Task<IEnumerable<Category>> Get()
     {
         return await _categoryCrud.GetAllCategories();
     }
 
-    //[Route("api/[controller]/one/{id:guid}/")]
-    //[Route("{id:guid}")]
-    //public async Task<Category> Get(Guid id)
-    //{
-    //    return await _categoryCrud.GetCategory(id);
-    //}
-
-
-
-
     [HttpPost]
     public async Task<IActionResult> Post(Category category)
     {
 
-        var result = await _categoryCrud.CreateCategory(category);
-        if (result) return Ok();
-        else return BadRequest();
+        var cust = HttpContext.Items["Customer"] as Customer;
+        if (cust is not null && cust.IsAdmin)
+        {
+            var result = await _categoryCrud.CreateCategory(category);
+            if (result) return Ok();
+            else return BadRequest();
+
+        }
+        else
+        {
+            return BadRequest(new { error = "Need admin priviledge to create category." });
+        }
+
+        return Ok();
     }
 
 
@@ -50,8 +54,22 @@ public class CategoryController : ControllerBase
     [HttpDelete]
     public async Task<IActionResult> Delete(Category category)
     {
-        var result = await _categoryCrud.DeleteCategory(category);
-        if (result) return Ok();
+        var cust = HttpContext.Items["Customer"] as Customer;
+        if (cust is not null && cust.IsAdmin)
+        {
+            var result = await _categoryCrud.DeleteCategory(category);
+            if (result) return Ok();
+            else return BadRequest();
+
+        }
+        else
+        {
+            return BadRequest(new { error = "Need admin priviledge to delete category." });
+        }
+
+
+        ////var result = await _categoryCrud.DeleteCategory(category);
+        ////if (result) return Ok();
         return BadRequest();
     }
